Add implicit C coercions in AST
This changes most of the test output and required to fix some code acting weirdly regarding to C types and casts.

+++ tests/apac/use_goto_for_return/apac_use_goto_for_return_exp.cpp
@@ -31,8 +31,8 @@ void h() {
 float i(float pi, float a) {
   float __res;
   {
-    float coef = 11.7f;
-    if (a < 0.) {
+    float coef = (float)11.7;
+    if ((double)a < 0.) {
       __res = coef * a;
       goto __exit;
     }
