+++ tests/loop/slide/loop_slide_exp.cpp
@@ -10,7 +10,7 @@ int main() {
   }
   for (int bj = 0; bj < 30 + (2 - 2); bj += 2) {
     for (int j = bj; j < min(30, bj + 2); j++) {
-      printf("%f\n", s[MINDEX1(32, j - 2)]);
+      printf("%f\n", (double)s[MINDEX1(32, j - 2)]);
     }
   }
   for (int bk = 0; bk < 15 + (1 - 2) * 3; bk += 1 * 3) {
