+++ tests/matrix/intro_calloc/matrix_intro_calloc_basic_exp.cpp
@@ -3,7 +3,7 @@
 
 int main() {
   const int N = 10;
-  int *p = (int *)CALLOC1(N, sizeof(int));
-  int *q = (int *)CALLOC1(N, sizeof(int));
+  int* p = (int*)CALLOC1((long unsigned int)N, sizeof(int));
+  int* q = (int*)CALLOC1((long unsigned int)N, sizeof(int));
   return 0;
 }
