+++ tests/matrix/intro_calloc/matrix_intro_calloc_exp.cpp
@@ -5,9 +5,9 @@ int* q;
 
 const int N = 10;
 
-void allocate() { q = (int *)CALLOC1(N, sizeof(int)); }
+void allocate() { q = (int*)CALLOC1((long unsigned int)N, sizeof(int)); }
 
 int main() {
-  int *p = (int *)CALLOC1(N, sizeof(int));
+  int* p = (int*)CALLOC1((long unsigned int)N, sizeof(int));
   return 0;
 }
