+++ tests/matrix/intro_malloc/matrix_intro_malloc_basic_doc_exp.cpp
@@ -3,6 +3,6 @@
 
 int main() {
   const int N = 10;
-  int* p = (int*)MALLOC1(N, sizeof(int));
+  int* p = (int*)MALLOC1((long unsigned int)N, sizeof(int));
   return 0;
 }
