+++ tests/matrix/intro_malloc/matrix_intro_malloc_doc_exp.cpp
@@ -3,5 +3,5 @@
 
 int main() {
   const int N = 10;
-  int* p = (int*)MALLOC1(N, sizeof(int));
+  int* p = (int*)MALLOC1((long unsigned int)N, sizeof(int));
 }
