+++ tests/matrix/stack_copy/matrix_stack_copy_doc_exp.cpp
@@ -5,7 +5,7 @@ int main() {
   for (int i = 0; i < 32; i++) {
     for (int j = 0; j < 32; j++) {
       for (int k = 0; k < 4; k++) {
-        s[MINDEX2(32, 32, i, j)] += k;
+        s[MINDEX2(32, 32, i, j)] += (float)k;
       }
     }
   }
